A few of the test files were moved into a more common xt/ location. To run those tests, it's easy to now just ```prove -lr xt```. Those tests have had their eval-wrapped use statements removed to be just plain use statements instead as these tests will not be run by any- one but the author.
A new test file, 00-report-prereqs.t now exists in the root of the
t/ directory to show a good amount of info during the test which
helps debug when an error appears from CPAN Testers reports or
from users.

The .gitignore file was altered a bit to ensure no files get tracked
after ```perl Makefile.PL```.

The MANIFEST was altered to indicate the xt tests described above.

The Makefile.PL file was altered to show any and all requirements,
even the trivial ones. This just shows a more-complete prereqs
report on a test run. A few creative greps over the code and tests
were made to ensure we have them all listed. The prereqs within
the xt/ directory are not listed as there isn't a spec for
DEVELOP_REQUIRES in the Makefile.PL and we're not tracking prereqs
with a cpanfile.

No actual code changes were made.

Overall, this is a pretty trivial change that should hopefully make
testing a bit easier and provide reports from users and smoke testers
when something blows up.